Rewrite Entitlements.plist to avoid newline issues.
Git can be configured globally to mangle newlines to CRLF.
Cookiecutter manages the git clone, so there's no easy way for us to
set the configuration at the repository level during the templating
process.

The macOS utility `codesign` fails with obscure, hard to debug errors
when presented with CRLF newlines in Entitlements.plist.